.agent/skills/SelfEvolution/requirements.txt
PyYAML>=6.0
google-generativeai>=0.3.0
python-dotenv>=1.0.0
